Automatically backed up by Learn

+++ index.js
@@ -34,6 +34,6 @@ function preserveDecimal (n){
 }
 
 function preserveDecimal (n){
-  return parseInt (n, 10);
+  return parseFloat (n, 10);
 }
